# GH-41361: [C++][Parquet] Optimize DelimitRecords by batch execution when max_rep_level > 1