@mapleFU 2024-04-24
On My M1Pro MacOS with Release(-O2):

After:

```
--------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------------------
RecordReaderSkipRecords/Repetition:0/BatchSize:1000                                   143900 ns       132519 ns         5141 bytes_per_second=35.9826G/s items_per_second=9.65901G/s
RecordReaderSkipRecords/Repetition:1/BatchSize:1000                                   730351 ns       709798 ns          986 bytes_per_second=3.57925G/s items_per_second=1.80333G/s
RecordReaderSkipRecords/Repetition:2/BatchSize:1000                                  2261865 ns      2072772 ns          342 bytes_per_second=1.307G/s items_per_second=617.531M/s
RecordReaderReadRecords/Repetition:0/BatchSize:1000/ReadDense:1                       145809 ns       142787 ns         4715 bytes_per_second=33.3949G/s items_per_second=8.96437G/s
RecordReaderReadRecords/Repetition:0/BatchSize:1000/ReadDense:0                       155365 ns       143254 ns         4970 bytes_per_second=33.2862G/s items_per_second=8.9352G/s
RecordReaderReadRecords/Repetition:1/BatchSize:1000/ReadDense:1                       822000 ns       812812 ns          847 bytes_per_second=3.12562G/s items_per_second=1.57478G/s
RecordReaderReadRecords/Repetition:1/BatchSize:1000/ReadDense:0                      4645842 ns      4490519 ns          156 bytes_per_second=579.335M/s items_per_second=285.045M/s
RecordReaderReadRecords/Repetition:2/BatchSize:1000/ReadDense:1                      2030276 ns      2015167 ns          347 bytes_per_second=1.34437G/s items_per_second=635.183M/s
RecordReaderReadRecords/Repetition:2/BatchSize:1000/ReadDense:0                      6398396 ns      6347872 ns          109 bytes_per_second=437.019M/s items_per_second=201.642M/s
RecordReaderReadAndSkipRecords/Repetition:0/BatchSize:10/LevelsPerPage:80000         3171922 ns      3074065 ns          230 bytes_per_second=1.55116G/s items_per_second=416.387M/s
RecordReaderReadAndSkipRecords/Repetition:0/BatchSize:1000/LevelsPerPage:80000        124805 ns       124232 ns         5412 bytes_per_second=38.3828G/s items_per_second=10.3033G/s
RecordReaderReadAndSkipRecords/Repetition:0/BatchSize:10000/LevelsPerPage:1000000    1538905 ns      1527279 ns          462 bytes_per_second=39.0267G/s items_per_second=10.4761G/s
RecordReaderReadAndSkipRecords/Repetition:1/BatchSize:10/LevelsPerPage:80000        13674456 ns     13436556 ns           54 bytes_per_second=193.615M/s items_per_second=95.2625M/s
RecordReaderReadAndSkipRecords/Repetition:1/BatchSize:1000/LevelsPerPage:80000       2558735 ns      2544208 ns          274 bytes_per_second=1022.52M/s items_per_second=503.104M/s
RecordReaderReadAndSkipRecords/Repetition:1/BatchSize:10000/LevelsPerPage:1000000   29944901 ns     29750043 ns           23 bytes_per_second=1091.78M/s items_per_second=537.814M/s
RecordReaderReadAndSkipRecords/Repetition:2/BatchSize:10/LevelsPerPage:80000        20192213 ns     17695683 ns           41 bytes_per_second=156.769M/s items_per_second=72.334M/s
RecordReaderReadAndSkipRecords/Repetition:2/BatchSize:100/LevelsPerPage:80000        5926492 ns      5729650 ns          123 bytes_per_second=484.173M/s items_per_second=223.399M/s
RecordReaderReadAndSkipRecords/Repetition:2/BatchSize:10000/LevelsPerPage:1000000   46863512 ns     46462071 ns           14 bytes_per_second=745.552M/s items_per_second=344.367M/s
```

Before:

```
RecordReaderSkipRecords/Repetition:0/BatchSize:1000                                   138199 ns       131139 ns         5460 bytes_per_second=36.3611G/s items_per_second=9.76062G/s
RecordReaderSkipRecords/Repetition:1/BatchSize:1000                                   710793 ns       703035 ns          992 bytes_per_second=3.61368G/s items_per_second=1.82068G/s
RecordReaderSkipRecords/Repetition:2/BatchSize:1000                                  5587717 ns      5362778 ns          135 bytes_per_second=517.295M/s items_per_second=238.682M/s
RecordReaderReadRecords/Repetition:0/BatchSize:1000/ReadDense:1                       172499 ns       149067 ns         5205 bytes_per_second=31.9881G/s items_per_second=8.58673G/s
RecordReaderReadRecords/Repetition:0/BatchSize:1000/ReadDense:0                       148973 ns       141769 ns         4672 bytes_per_second=33.6349G/s items_per_second=9.0288G/s
RecordReaderReadRecords/Repetition:1/BatchSize:1000/ReadDense:1                       824737 ns       808909 ns          875 bytes_per_second=3.1407G/s items_per_second=1.58238G/s
RecordReaderReadRecords/Repetition:1/BatchSize:1000/ReadDense:0                      4577248 ns      4525734 ns          154 bytes_per_second=574.827M/s items_per_second=282.827M/s
RecordReaderReadRecords/Repetition:2/BatchSize:1000/ReadDense:1                      5774188 ns      5430985 ns          131 bytes_per_second=510.799M/s items_per_second=235.685M/s
RecordReaderReadRecords/Repetition:2/BatchSize:1000/ReadDense:0                      9687532 ns      9634611 ns           72 bytes_per_second=287.935M/s items_per_second=132.854M/s
RecordReaderReadAndSkipRecords/Repetition:0/BatchSize:10/LevelsPerPage:80000         3091099 ns      3071394 ns          226 bytes_per_second=1.55251G/s items_per_second=416.749M/s
RecordReaderReadAndSkipRecords/Repetition:0/BatchSize:1000/LevelsPerPage:80000        134295 ns       131932 ns         5225 bytes_per_second=36.1427G/s items_per_second=9.70198G/s
RecordReaderReadAndSkipRecords/Repetition:0/BatchSize:10000/LevelsPerPage:1000000    1579842 ns      1527460 ns          452 bytes_per_second=39.0221G/s items_per_second=10.4749G/s
RecordReaderReadAndSkipRecords/Repetition:1/BatchSize:10/LevelsPerPage:80000        13334181 ns     13274907 ns           54 bytes_per_second=195.972M/s items_per_second=96.4225M/s
RecordReaderReadAndSkipRecords/Repetition:1/BatchSize:1000/LevelsPerPage:80000       2554745 ns      2539457 ns          276 bytes_per_second=1024.44M/s items_per_second=504.045M/s
RecordReaderReadAndSkipRecords/Repetition:1/BatchSize:10000/LevelsPerPage:1000000   30177870 ns     29929000 ns           23 bytes_per_second=1085.25M/s items_per_second=534.599M/s
RecordReaderReadAndSkipRecords/Repetition:2/BatchSize:10/LevelsPerPage:80000        18526031 ns     18415737 ns           38 bytes_per_second=150.64M/s items_per_second=69.5058M/s
RecordReaderReadAndSkipRecords/Repetition:2/BatchSize:100/LevelsPerPage:80000        9389062 ns      8528720 ns           82 bytes_per_second=325.27M/s items_per_second=150.081M/s
RecordReaderReadAndSkipRecords/Repetition:2/BatchSize:10000/LevelsPerPage:1000000   88131281 ns     87666250 ns            8 bytes_per_second=395.134M/s items_per_second=182.51M/s
```